npu_camera: combine main sources.
