+++ src/img_sub.cpp
@@ -66,3 +66,12 @@ void CamSubscriber_::process_image(const sensor_msgs::msg::Image::SharedPtr msg)
   if (cv::waitKey(1) == 'q')
     exit(0);
 }
+
+int main(int argc, char * argv[])
+{
+  rclcpp::init(argc, argv);
+  auto node = std::make_shared<CamSubscriber_>();
+  rclcpp::spin(node);
+  rclcpp::shutdown();
+  return 0;
+}
