--- src/sub_main.cpp
@@ -1,10 +1,0 @@
-#include "img_sub.h"
-
-int main(int argc, char * argv[])
-{
-  rclcpp::init(argc, argv);
-  auto node = std::make_shared<CamSubscriber_>();
-  rclcpp::spin(node);
-  rclcpp::shutdown();
-  return 0;
-}
